Package re structure
