+++ src/react-local-file-system/components/ApplyContextMenu.jsx
@@ -1,0 +1,55 @@
+import React, { useState } from "react";
+import { Menu, MenuItem } from "@mui/material";
+
+export default function ApplyContextMenu({ children, items }) {
+    const [contextMenu, setContextMenu] = useState(null);
+
+    const handleContextMenu = (event) => {
+        event.preventDefault();
+        setContextMenu(
+            contextMenu === null
+                ? {
+                      mouseX: event.clientX,
+                      mouseY: event.clientY,
+                  }
+                : // repeated contextmenu when it is already open closes it with Chrome 84 on Ubuntu
+                  // Other native context menus might behave different.
+                  // With this behavior we prevent contextmenu from the backdrop to re-locale existing context menus.
+                  null
+        );
+    };
+
+    const handleClose = () => {
+        setContextMenu(null);
+    };
+
+    return (
+        <>
+            <div onContextMenu={handleContextMenu} style={{ cursor: "context-menu" }}>
+                {children}
+                <Menu
+                    open={contextMenu !== null}
+                    onClose={handleClose}
+                    anchorReference="anchorPosition"
+                    anchorPosition={
+                        contextMenu !== null ? { top: contextMenu.mouseY, left: contextMenu.mouseX } : undefined
+                    }
+                >
+                    {items.map((item) => {
+                        return (
+                            <MenuItem
+                                key={"local_file_system_menu_item_key_" + item.name}
+                                onClick={(event) => {
+                                    handleClose();
+                                    item.handler(event);
+                                }}
+                            >
+                                {item.name}
+                            </MenuItem>
+                        );
+                    })}
+                </Menu>
+            </div>
+        </>
+    );
+}
