+++ src/react-local-file-system/contexts/CurFolderContext.js
@@ -1,0 +1,5 @@
+import { createContext } from "react";
+
+const CurFolderContext = createContext();
+
+export default CurFolderContext;
