+++ src/react-local-file-system/contexts/DragContext.js
@@ -1,0 +1,5 @@
+import { createContext } from "react";
+
+const DragContext = createContext();
+
+export default DragContext;
